Address document title hook review comment
