+++ airflow-core/src/airflow/ui/src/utils/documentTitleContext.ts
@@ -16,14 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+import { createContext } from "react";
 
-export { capitalize } from "./capitalize";
-export { getDuration, renderDuration } from "./datetimeUtils";
-export { createErrorToaster, getErrorStatus } from "./errorHandling";
-export { getMetaKey } from "./getMetaKey";
-export { useContainerWidth } from "./useContainerWidth";
-export { useDocumentTitle } from "./useDocumentTitle";
-export { DocumentTitleProvider } from "./useDocumentTitleProvider";
-export { useFiltersHandler, type FilterableSearchParamsKeys } from "./useFiltersHandler";
-export * from "./query";
-export { STATE_PRIORITY, sortStateEntries } from "./stateUtils";
+export const DocumentTitleContext = createContext<((pageTitle: string | null) => void) | undefined>(
+  undefined,
+);
